Refactored ArenaBrawl to arena_brawl package and HypixelArcade to arcade package
