+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/arena_brawl/HypixelArenaBrawl.java
@@ -1,4 +1,6 @@
-package io.github.hypixel_api_wrapper.wrapper.games;
+package io.github.hypixel_api_wrapper.wrapper.games.arena_brawl;
+
+import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
 
 public class HypixelArenaBrawl implements HypixelGame {
     @Override
